+++ src/lotus_json/signed_message.rs
@@ -52,11 +52,11 @@ impl HasLotusJson for SignedMessage {
                     "To": "f00",
                     "Value": "0",
                     "Version": 0,
-                    "CID": {
-                        "/": "bafy2bzaced3xdk2uf6azekyxgcttujvy3fzyeqmibtpjf2fxcpfdx2zcx4s3g"
-                    },
                 },
-                "Signature": {"Type": 2, "Data": "aGVsbG8gd29ybGQh"}
+                "Signature": {"Type": 2, "Data": "aGVsbG8gd29ybGQh"},
+                "CID": {
+                    "/": "bafy2bzaced3xdk2uf6azekyxgcttujvy3fzyeqmibtpjf2fxcpfdx2zcx4s3g"
+                },
             }),
             SignedMessage {
                 message: crate::shim::message::Message::default(),
@@ -69,11 +69,12 @@ impl HasLotusJson for SignedMessage {
     }
 
     fn into_lotus_json(self) -> Self::LotusJson {
+        let cid = self.cid().ok();
         let Self { message, signature } = self;
         Self::LotusJson {
             message,
             signature,
-            cid: None, // See notes on Message
+            cid,
         }
     }